Added authorisation for clients

+++ src/models/client.py
@@ -14,9 +14,9 @@ class Client(BaseModel):
     cui: str
     user_id: int | None = None
 
-    def add(self):
+    def add(self, user_id: int):
         query = "INSERT INTO clients(name, address, contact_person, country, phone_number, onrc_no, cui, user_id) VALUES (?,?,?,?,?,?,?,?)"
-        self.user_id = 0
+        self.user_id = user_id
         data = (self.name, self.address, self.contact_person, self.country, self.phone_number, self.onrc_no, self.cui, self.user_id)
         try: 
             res: sqlite3.Cursor = db.execute_query(query=query, params=data)
@@ -26,41 +26,46 @@ class Client(BaseModel):
             raise HTTPException(500,e.args[0])
 
     def update(self, client_id, user_id):
-        query = "UPDATE clients	SET name = ?, address = ?, contact_person = ?, country = ?, phone_number = ?, onrc_no = ?, cui = ?	WHERE id = ?"
-        self.user_id = user_id
-        self.id = client_id
-        data = (self.name, self.address, self.contact_person, self.country, self.phone_number, self.onrc_no, self.cui, self.id)
+        check_permissions: bool = check_user_id(client_id=client_id, user_id=user_id)
+        if check_permissions:
+            query = "UPDATE clients	SET name = ?, address = ?, contact_person = ?, country = ?, phone_number = ?, onrc_no = ?, cui = ?	WHERE id = ?"
+            self.user_id = user_id
+            self.id = client_id
+            data = (self.name, self.address, self.contact_person, self.country, self.phone_number, self.onrc_no, self.cui, self.id)
+            try:
+                _: sqlite3.Cursor = db.execute_query(query=query, params=data)
+                return self
+            except sqlite3.IntegrityError as e:
+                raise HTTPException(500,e.args[0])
+    
+def show_user_client(client_id: int, user_id: int):
+    check_permissions: bool = check_user_id(client_id=client_id, user_id=user_id)
+    if check_permissions:
+        query = "SELECT * FROM clients WHERE id == ? AND user_id == ?"
+        data = (client_id, user_id)
         try: 
-            _: sqlite3.Cursor = db.execute_query(query=query, params=data)
-            return self
+            res: sqlite3.Cursor = db.execute_query(query=query, params=data)
+            row: tuple = res.fetchone()
+            client: Client = Client(id=int(row[0]),
+                            name=row[1],
+                            address=row[2],
+                            contact_person=row[3],
+                            country=row[4],
+                            phone_number=row[5],
+                            onrc_no=row[6],
+                            cui=row[7],
+                            user_id=int(row[8])
+            )
+            return client
         except sqlite3.IntegrityError as e:
             raise HTTPException(500,e.args[0])
-    
-def show_client(client_id: int):
-    query = "SELECT * FROM clients WHERE id == ?"
-    data = (client_id,)
+
+def list_user_clients(user_id: int):
+    client_list: list[Client] = []
+    query = "SELECT * FROM clients where user_id == ?"
+    data= (user_id,)
     try: 
         res: sqlite3.Cursor = db.execute_query(query=query, params=data)
-        row: tuple = res.fetchone()
-        client: Client = Client(id=int(row[0]),
-                        name=row[1],
-                        address=row[2],
-                        contact_person=row[3],
-                        country=row[4],
-                        phone_number=row[5],
-                        onrc_no=row[6],
-                        cui=row[7],
-                        user_id=int(row[8])
-        )
-        return client
-    except sqlite3.IntegrityError as e:
-        raise HTTPException(500,e.args[0])
-
-def list_clients():
-    client_list: list[Client] = []
-    query = "SELECT * FROM clients"
-    try: 
-        res: sqlite3.Cursor = db.execute_query(query=query)
         rows: list[tuple] = res.fetchall()
         for row in rows:
             client: Client = Client(id=int(row[0]),
@@ -77,12 +82,30 @@ def list_clients():
     except sqlite3.OperationalError as e:
         raise HTTPException(500,e.args[0])
 
-def delete_client(client_id: int):
-    query = "DELETE FROM clients WHERE id == ?"
+def delete_client(client_id: int, user_id: int):
+    check_permissions: bool = check_user_id(client_id=client_id, user_id=user_id)
+    if check_permissions:
+        query = "DELETE FROM clients WHERE id == ? AND user_id == ?"
+        data = (client_id, user_id)
+        try: 
+            _: sqlite3.Cursor = db.execute_query(query=query, params=data)
+            return True
+        except sqlite3.IntegrityError as e:
+            raise HTTPException(500,e.args[0])
+
+def check_user_id(client_id: int, user_id: int):
+    query = "SELECT user_id FROM clients WHERE id == ?"
     data = (client_id,)
     try: 
-        _: sqlite3.Cursor = db.execute_query(query=query, params=data)
-        return True
-    except sqlite3.IntegrityError as e:
+        res: sqlite3.Cursor = db.execute_query(query=query, params=data)
+        row: tuple = res.fetchone()
+        if row:
+            client_user_id:int= int(row[0])
+        else:
+            raise HTTPException(status_code=403,detail="No such client")
+    except sqlite3.Error as e:
         raise HTTPException(500,e.args[0])
-
+    if client_user_id == user_id:
+        return True
+    else:
+        raise HTTPException(status_code=403, detail="You are not allowed to perform this action")
